+++ src/App.js
@@ -1,21 +1,21 @@
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { CookiesProvider } from "react-cookie";
 import { RecoilRoot } from "recoil";
-import ScrollRestoration from "pages/Skeleton/ScrollRestoration";
-import VirtualKeyboardDetector from "pages/Skeleton/VirtualKeyboardDetector";
-import ChannelTalkProvider from "pages/Skeleton/ChannelTalkProvider";
-import GoogleAnalyticsProvier from "pages/Skeleton/GoogleAnalyticsProvier";
-import I18nextProvider from "lang/I18nextProvider";
-import AlertProvider from "pages/Skeleton/AlertProvider";
-import CSSVariablesProvider from "pages/Skeleton/CSSVariablesProvider";
-import Skeleton from "pages/Skeleton/Skeleton";
+import ScrollRestoration from "components/Skeleton/ScrollRestoration";
+import VirtualKeyboardDetector from "components/Skeleton/VirtualKeyboardDetector";
+import ChannelTalkProvider from "components/Skeleton/ChannelTalkProvider";
+import GoogleAnalyticsProvier from "components/Skeleton/GoogleAnalyticsProvier";
+import I18nextProvider from "components/Skeleton/I18nextProvider";
+import AlertProvider from "components/Skeleton/AlertProvider";
+import CSSVariablesProvider from "components/Skeleton/CSSVariablesProvider";
+import Skeleton from "components/Skeleton/Skeleton";
 
 import Login from "pages/Login/Login";
 import LoginFail from "pages/Login/LoginFail";
 import Logout from "pages/Login/Logout";
 import Home from "pages/Home";
 import Search from "pages/Search/Search";
-import AddRoom from "pages/AddRoom/AddRoom";
+import AddRoom from "pages/Addroom/AddRoom";
 import Myroom from "pages/Myroom/Myroom";
 import Mypage from "pages/Mypage/Mypage";
 import WrapChat from "pages/Chatting/WrapChat";
