+++ src/components/Skeleton/CSSVariablesProvider.tsx
