+++ src/components/Skeleton/Footer.tsx
